# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Game Store API",
    "description": "A complete game store API with customers, products, and invoices management",
    "version": "1.0.0",
    "contact": {
      "name": "Game Store API Support",
      "email": "support@gamestore.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/customers": {
      "get": {
        "summary": "List all customers",
        "description": "Get a list of all customers",
        "tags": ["Customers"],
        "responses": {
          "200": {
            "description": "List of customers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new customer",
        "description": "Create a new customer in the store",
        "tags": ["Customers"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "409": {
            "description": "Customer with this email already exists"
          }
        }
      }
    },
    "/api/customers/{id}": {
      "get": {
        "summary": "Get customer by ID",
        "description": "Get a specific customer by their ID",
        "tags": ["Customers"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Customer ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "400": {
            "description": "ID is required"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      },
      "put": {
        "summary": "Update customer",
        "description": "Update an existing customer",
        "tags": ["Customers"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Customer ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Customer not found"
          },
          "409": {
            "description": "Customer with this email already exists"
          }
        }
      },
      "delete": {
        "summary": "Delete customer",
        "description": "Delete a customer by ID",
        "tags": ["Customers"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Customer ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Customer deleted successfully"
          },
          "400": {
            "description": "ID is required"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      }
    },
    "/api/products": {
      "get": {
        "summary": "List all products",
        "description": "Get a list of all products with optional filtering",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "Filter only active products"
          }
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new product",
        "description": "Create a new product in the store",
        "tags": ["Products"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "409": {
            "description": "Product with this name already exists"
          }
        }
      }
    },
    "/api/products/category": {
      "get": {
        "summary": "Get products by category",
        "description": "Get products filtered by category",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product category"
          }
        ],
        "responses": {
          "200": {
            "description": "List of products in category",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Category is required"
          }
        }
      }
    },
    "/api/products/platform": {
      "get": {
        "summary": "Get products by platform",
        "description": "Get products filtered by platform",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product platform"
          }
        ],
        "responses": {
          "200": {
            "description": "List of products for platform",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Platform is required"
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "Get product by ID",
        "description": "Get a specific product by its ID",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Product found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "ID is required"
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "put": {
        "summary": "Update product",
        "description": "Update an existing product",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Product not found"
          },
          "409": {
            "description": "Product with this name already exists"
          }
        }
      },
      "delete": {
        "summary": "Delete product",
        "description": "Delete a product by ID",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted successfully"
          },
          "400": {
            "description": "ID is required"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
    "/api/products/{id}/stock": {
      "patch": {
        "summary": "Update product stock",
        "description": "Update the stock quantity of a product",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStockRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stock updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
    "/api/invoices": {
      "get": {
        "summary": "List all invoices",
        "description": "Get a list of all invoices with optional filtering",
        "tags": ["Invoices"],
        "parameters": [
          {
            "name": "customer_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter by customer ID"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["pending", "paid", "cancelled"]
            },
            "description": "Filter by invoice status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of invoices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invoice"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new invoice",
        "description": "Create a new invoice in the store",
        "tags": ["Invoices"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoiceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invoice created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      }
    },
    "/api/invoices/period": {
      "get": {
        "summary": "Get invoices by period",
        "description": "Get invoices filtered by date period",
        "tags": ["Invoices"],
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Start date (YYYY-MM-DD)"
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "End date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of invoices in period",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invoice"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid date parameters"
          }
        }
      }
    },
    "/api/invoices/total": {
      "get": {
        "summary": "Get total sales",
        "description": "Get total sales amount with optional period filter",
        "tags": ["Invoices"],
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Start date (YYYY-MM-DD)"
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "End date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Total sales amount",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "number",
                      "format": "float",
                      "example": 1500.50
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/invoices/{id}": {
      "get": {
        "summary": "Get invoice by ID",
        "description": "Get a specific invoice by its ID",
        "tags": ["Invoices"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Invoice ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "400": {
            "description": "ID is required"
          },
          "404": {
            "description": "Invoice not found"
          }
        }
      },
      "put": {
        "summary": "Update invoice",
        "description": "Update an existing invoice",
        "tags": ["Invoices"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Invoice ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvoiceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invoice updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Invoice not found"
          }
        }
      },
      "delete": {
        "summary": "Delete invoice",
        "description": "Delete an invoice by ID",
        "tags": ["Invoices"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Invoice ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Invoice deleted successfully"
          },
          "400": {
            "description": "ID is required"
          },
          "404": {
            "description": "Invoice not found"
          }
        }
      }
    },
    "/api/invoices/{id}/paid": {
      "patch": {
        "summary": "Mark invoice as paid",
        "description": "Mark an invoice as paid",
        "tags": ["Invoices"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Invoice ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice marked as paid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "400": {
            "description": "ID is required"
          },
          "404": {
            "description": "Invoice not found"
          }
        }
      }
    },
    "/api/invoices/{id}/cancel": {
      "patch": {
        "summary": "Cancel invoice",
        "description": "Cancel an invoice",
        "tags": ["Invoices"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Invoice ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "400": {
            "description": "ID is required"
          },
          "404": {
            "description": "Invoice not found"
          }
        }
      }
    },
    "/api/invoices/{id}/items": {
      "post": {
        "summary": "Add item to invoice",
        "description": "Add an item to an invoice",
        "tags": ["Invoices"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Invoice ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Invoice not found"
          }
        }
      },
      "put": {
        "summary": "Update invoice item",
        "description": "Update the quantity of an item in an invoice",
        "tags": ["Invoices"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Invoice ID"
          },
          {
            "name": "product_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Invoice or item not found"
          }
        }
      },
      "delete": {
        "summary": "Remove invoice item",
        "description": "Remove an item from an invoice",
        "tags": ["Invoices"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Invoice ID"
          },
          {
            "name": "product_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Item removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "404": {
            "description": "Invoice or item not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique customer identifier"
          },
          "name": {
            "type": "string",
            "description": "Customer full name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Customer email address"
          },
          "phone": {
            "type": "string",
            "description": "Customer phone number"
          },
          "address": {
            "type": "string",
            "description": "Customer address"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Customer creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Customer last update timestamp"
          }
        },
        "required": ["id", "name", "email", "created_at", "updated_at"]
      },
      "CreateCustomerRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Customer full name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Customer email address"
          },
          "phone": {
            "type": "string",
            "description": "Customer phone number"
          },
          "address": {
            "type": "string",
            "description": "Customer address"
          }
        },
        "required": ["name", "email"]
      },
      "UpdateCustomerRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Customer full name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Customer email address"
          },
          "phone": {
            "type": "string",
            "description": "Customer phone number"
          },
          "address": {
            "type": "string",
            "description": "Customer address"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique product identifier"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "description": {
            "type": "string",
            "description": "Product description"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "Product price"
          },
          "category": {
            "type": "string",
            "description": "Product category"
          },
          "platform": {
            "type": "string",
            "description": "Product platform"
          },
          "stock": {
            "type": "integer",
            "description": "Available stock quantity"
          },
          "active": {
            "type": "boolean",
            "description": "Product active status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Product creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Product last update timestamp"
          }
        },
        "required": ["id", "name", "description", "price", "created_at", "updated_at"]
      },
      "CreateProductRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "description": {
            "type": "string",
            "description": "Product description"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "Product price"
          },
          "category": {
            "type": "string",
            "description": "Product category"
          },
          "platform": {
            "type": "string",
            "description": "Product platform"
          },
          "stock": {
            "type": "integer",
            "description": "Available stock quantity"
          }
        },
        "required": ["name", "description", "price"]
      },
      "UpdateProductRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "description": {
            "type": "string",
            "description": "Product description"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "Product price"
          },
          "category": {
            "type": "string",
            "description": "Product category"
          },
          "platform": {
            "type": "string",
            "description": "Product platform"
          },
          "stock": {
            "type": "integer",
            "description": "Available stock quantity"
          },
          "active": {
            "type": "boolean",
            "description": "Product active status"
          }
        }
      },
      "UpdateStockRequest": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "integer",
            "description": "Quantity to add/subtract from stock"
          }
        },
        "required": ["quantity"]
      },
...
      "CreateInvoiceRequest": {
        "type": "object",
        "properties": {
          "customer_id": {
            "type": "string",
            "description": "Customer ID"
          },
          "customer_name": {
            "type": "string",
            "description": "Customer name"
          },
          "payment_method": {
            "type": "string",
            "enum": ["cash", "card", "pix", "boleto"],
            "description": "Payment method"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes"
          }
        },
        "required": ["customer_id", "customer_name", "payment_method"]
      },
...
      "UpdateItemRequest": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "integer",
            "description": "New item quantity"
          }
        },
        "required": ["quantity"]
      }
    }
  },
...
}